Page about + Avatar component done
