+++ front/src/Pages/Homepage/Homepage.css
@@ -1,0 +1,9 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer base {
+  body {
+    background-color:#1b1b1c;
+  }
+}
