--- front/src/front/Pages/Homepage/Homepage.css
@@ -1,9 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer base {
-  body {
-    background-color:#1b1b1c;
-  }
-}
